# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/loyalty_programs:
    get:
      tags:
        - LoyaltyPrograms
      summary: List LoyaltyPrograms
      description: Get a paginated list of LoyaltyPrograms with optional filtering and sorting
      operationId: listLoyaltyPrograms
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyProgramList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - LoyaltyPrograms
      summary: Create LoyaltyProgram
      description: Create a new LoyaltyProgram
      operationId: createLoyaltyProgram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoyaltyProgramRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyProgram'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/loyalty_point_entries:
    get:
      tags:
        - LoyaltyPointEntries
      summary: List LoyaltyPointEntries
      description: Get a paginated list of LoyaltyPointEntries with optional filtering and sorting
      operationId: listLoyaltyPointEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyPointEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - LoyaltyPointEntries
      summary: Create LoyaltyPointEntry
      description: Create a new LoyaltyPointEntry
      operationId: createLoyaltyPointEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoyaltyPointEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyPointEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    LoyaltyProgram:
      type: object
      required:
        - id
        - company_id
        - program_name
        - program_type
        - collection_factor
        - conversion_factor
        - from_date
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        program_name:
          type: string
          maxLength: 140
        program_type:
          type: string
        collection_factor:
          type: string
        conversion_factor:
          type: string
        expiry_duration_days:
          nullable: true
          type: integer
          format: int32
        from_date:
          type: string
          format: date-time
        to_date:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        metadata:
          type: object
...
    CreateLoyaltyProgramRequest:
      type: object
      required:
        - company_id
        - program_name
        - program_type
        - collection_factor
        - conversion_factor
        - from_date
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        program_name:
          type: string
          maxLength: 140
        program_type:
          type: string
        collection_factor:
          type: string
        conversion_factor:
          type: string
        expiry_duration_days:
          nullable: true
          type: integer
          format: int32
        from_date:
          type: string
          format: date-time
        to_date:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        metadata:
          type: object
...
    LoyaltyProgramList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LoyaltyProgram'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    LoyaltyPointEntry:
      type: object
      required:
        - id
        - company_id
        - loyalty_program_id
        - customer_id
        - entry_type
        - points
        - purchase_amount
        - source_type
        - source_id
        - posting_date
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        loyalty_program_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        entry_type:
          type: string
        points:
          type: string
        purchase_amount:
          type: string
        source_type:
          type: string
          maxLength: 40
        source_id:
          type: string
          format: uuid
        posting_date:
          type: string
          format: date-time
        expiry_date:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateLoyaltyPointEntryRequest:
      type: object
      required:
        - company_id
        - loyalty_program_id
        - customer_id
        - entry_type
        - points
        - purchase_amount
        - source_type
        - source_id
        - posting_date
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        loyalty_program_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        entry_type:
          type: string
        points:
          type: string
        purchase_amount:
          type: string
        source_type:
          type: string
          maxLength: 40
        source_id:
          type: string
          format: uuid
        posting_date:
          type: string
          format: date-time
        expiry_date:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    LoyaltyPointEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LoyaltyPointEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []