# src/presentation/http/loyalty_program_handler.rs
//! LoyaltyProgram REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{LoyaltyProgramService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateLoyaltyProgramDto, UpdateLoyaltyProgramDto, PatchLoyaltyProgramDto, LoyaltyProgramResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum LoyaltyProgramError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for LoyaltyProgramError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for LoyaltyProgramError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "LOYALTYPROGRAM_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "LOYALTYPROGRAM_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "LOYALTYPROGRAM_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "LOYALTYPROGRAM_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for LoyaltyProgram.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /loyalty_programs | List with pagination |
/// | POST | /loyalty_programs | Create new |
/// | GET | /loyalty_programs/:id | Get by ID |
/// | PUT | /loyalty_programs/:id | Full update |
/// | PATCH | /loyalty_programs/:id | Partial update |
/// | DELETE | /loyalty_programs/:id | Soft delete |
/// | POST | /loyalty_programs/bulk | Bulk create |
/// | POST | /loyalty_programs/upsert | Upsert |
/// | GET | /loyalty_programs/trash | List deleted |
/// | POST | /loyalty_programs/:id/restore | Restore |
/// | DELETE | /loyalty_programs/empty | Empty trash |
/// | GET | /loyalty_programs/:id/deleted | Get deleted by ID |
/// | DELETE | /loyalty_programs/trash/:id | Permanent delete from trash |
/// | GET | /loyalty_programs/count | Count active entities |
/// | GET | /loyalty_programs/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(LoyaltyProgramService::with_repository(repository));
/// let router = create_loyalty_program_routes(service);
/// ```
pub fn create_loyalty_program_routes(service: Arc<LoyaltyProgramService>) -> Router {
    BackboneCrudHandler::<LoyaltyProgramService, LoyaltyProgram, CreateLoyaltyProgramDto, UpdateLoyaltyProgramDto, LoyaltyProgramResponseDto>::routes(
        service,
        "/loyalty_programs",
    )
}

/// Create Axum router with only the read (GET) endpoints for LoyaltyProgram.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_loyalty_program_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_loyalty_program_read_routes(service: Arc<LoyaltyProgramService>) -> Router {
    BackboneCrudHandler::<LoyaltyProgramService, LoyaltyProgram, CreateLoyaltyProgramDto, UpdateLoyaltyProgramDto, LoyaltyProgramResponseDto>::read_routes(
        service,
        "/loyalty_programs",
    )
}

/// Create Axum router with only the write (mutation) endpoints for LoyaltyProgram.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_loyalty_program_write_routes(service: Arc<LoyaltyProgramService>) -> Router {
    BackboneCrudHandler::<LoyaltyProgramService, LoyaltyProgram, CreateLoyaltyProgramDto, UpdateLoyaltyProgramDto, LoyaltyProgramResponseDto>::write_routes(
        service,
        "/loyalty_programs",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_loyalty_program_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<LoyaltyProgramService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_loyalty_program_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}